Enable scale to zero without an invocation
The existing scale to zero code was such that Prom stats needed to be available in order for faas-idler to scale a function to zero.  This meant that on deploy a function would need to be invoked in order to generate Prom stats and consequently scale the function to zero.

This change will remove the need for an invocation to occur in order to scale to zero by allowing entry to the metric generating code when there is no Prom data but the invocation count is zero.  The code to create a default value of zero in the metric map has been pulled up out of the data raging loop to enable it to be set in situations where there is no data.

Signed-off-by: Richard Gee <richard@technologee.co.uk>